Made separate test class. My test still doesn't work but now it's separate.
