+++ test/CleanWaterTests.java
@@ -1,7 +1,5 @@
 package test;
 import controller.HistReportController;
-import controller.MainScreenController;
-import fxapp.MainFXApplication;
 import model.*;
 
 import java.io.IOException;
@@ -27,7 +25,6 @@ public class CleanWaterTests {
     private HistoricalData virusData;
     private HistoricalData contaminantData;
     private List<Report> streamList;
-    private MainScreenController mainController;
 
     private class CleanOutputStream extends OutputStream {
 
@@ -242,22 +239,5 @@ public class CleanWaterTests {
         assertTrue(r1.wasReportCreatedFirst(r2));
         assertTrue(r2.wasReportCreatedFirst(r3));
         assertTrue(r3.wasReportCreatedFirst(r4));
-    }
-
-    /**
-     * Tests the method MainScreenController.onSubmitQualityReport
-     */
-    @Test(timeout=TIMEOUT)
-    public void testOnSubmitQualityReport() {
-        User user = new User("user", null, PermissionLevel.USER);
-        User worker = new User("worker", null, PermissionLevel.WORKER);
-        User manager = new User("worker", null, PermissionLevel.MANAGER);
-        User admin = new User("admin", null, PermissionLevel.ADMIN);
-        MainFXApplication testMain = new MainFXApplication();
-        MainScreenController testMainController = new MainScreenController();
-        testMainController.registerMainApp(testMain);
-        testMain.setActiveUser(manager);
-        testMainController.onSubmitQualityReport();
-        assertTrue(testMain.getActiveScene().getTitle() == "Cleanwater - Submit quality report");
     }
 }
